# Conflicts:
#	Source/Runtime/Private/Core/Task.cpp
#	Source/Runtime/Private/Platform/Android/PlatformAndroidThread.cpp
#	Source/Runtime/Private/Platform/PlatformBaseThread.cpp
#	Source/Runtime/Private/Platform/Posix/PlatformPosixThread.cpp
#	Source/Runtime/Private/Platform/Windows/PlatformWinThread.cpp
#	Source/Runtime/Public/Core/Task.h
#	Source/Runtime/Public/Platform/Android/PlatformAndroidThread.h
#	Source/Runtime/Public/Platform/PlatformThread.h
#	Source/Runtime/Public/Platform/Posix/PlatformPosixThread.h
#	Source/Runtime/Public/Platform/Windows/PlatformWinThread.h
#	cmake-vs2019-Win64-All.bat
#	cmake-vs2019-Win64.bat

+++ Source/Runtime/Public/Core/DateTime.h
@@ -88,7 +88,7 @@ public:
     int32_t             Day() const { return ToDate().day; }
     double              JulianDay() const { return (double)(1721425.5 + ticks / Timespan::TicksPerDay); }
     double              ModifiedJulianDay() const { return (JulianDay() - 2400000.5); }
-    DayOfWeek::Enum     DayOfWeek() const { return static_cast<DayOfWeek::Enum>((ticks / Timespan::TicksPerDay) % 7); }
+    DayOfWeek::Enum     DayOfWeek() const { return static_cast<DayOfWeek::Enum>((int)(ticks / Timespan::TicksPerDay) % 7); }
     int32_t             DayOfYear() const;
     int32_t             Hour() const { return (int32_t)((ticks / Timespan::TicksPerHour) % 24); }
     int32_t             Hour12() const;
